requirements.txt
Django==3.2.16
drf-nested-routers==0.93.4
djangorestframework==3.14.0
django-filter==21.1